requirements.txt
Flask_WTF==0.14.2
Flask==1.0.2
Flask_SQLAlchemy==2.3.2
python-twitter==3.4.1
newsapi-python==0.2.3
pyEX==0.1.0
validators
